+++ client/src/main/java/org/eclipse/hono/client/ClientErrorException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2017 Contributors to the Eclipse Foundation
+ * Copyright (c) 2016, 2020 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -29,7 +29,18 @@ public class ClientErrorException extends ServiceInvocationException {
      * @throws IllegalArgumentException if the code is not &ge; 400 and &lt; 500.
      */
     public ClientErrorException(final int errorCode) {
-        this(errorCode, null, null);
+        this(null, errorCode, null, null);
+    }
+
+    /**
+     * Creates a new exception for a tenant and a client error code.
+     *
+     * @param tenant The tenant that the exception occurred in the scope of or {@code null} if unknown.
+     * @param errorCode The code representing the erroneous outcome.
+     * @throws IllegalArgumentException if the code is not &ge; 400 and &lt; 500.
+     */
+    public ClientErrorException(final String tenant, final int errorCode) {
+        this(tenant, errorCode, null, null);
     }
 
     /**
@@ -40,7 +51,19 @@ public class ClientErrorException extends ServiceInvocationException {
      * @throws IllegalArgumentException if the code is not &ge; 400 and &lt; 500.
      */
     public ClientErrorException(final int errorCode, final String msg) {
-        this(errorCode, msg, null);
+        this(null, errorCode, msg, null);
+    }
+
+    /**
+     * Creates a new exception for a tenant, a client error code and a detail message.
+     *
+     * @param tenant The tenant that the exception occurred in the scope of or {@code null} if unknown.
+     * @param errorCode The code representing the erroneous outcome.
+     * @param msg The detail message.
+     * @throws IllegalArgumentException if the code is not &ge; 400 and &lt; 500.
+     */
+    public ClientErrorException(final String tenant, final int errorCode, final String msg) {
+        this(tenant, errorCode, msg, null);
     }
 
     /**
@@ -51,7 +74,19 @@ public class ClientErrorException extends ServiceInvocationException {
      * @throws IllegalArgumentException if the code is not &ge; 400 and &lt; 500.
      */
     public ClientErrorException(final int errorCode, final Throwable cause) {
-        this(errorCode, null, cause);
+        this(null, errorCode, null, cause);
+    }
+
+    /**
+     * Creates a new exception for a client error code and a root cause.
+     *
+     * @param tenant The tenant that the exception occurred in the scope of or {@code null} if unknown.
+     * @param errorCode The code representing the erroneous outcome.
+     * @param cause The root cause.
+     * @throws IllegalArgumentException if the code is not &ge; 400 and &lt; 500.
+     */
+    public ClientErrorException(final String tenant, final int errorCode, final Throwable cause) {
+        this(tenant, errorCode, null, cause);
     }
 
     /**
@@ -63,10 +98,22 @@ public class ClientErrorException extends ServiceInvocationException {
      * @throws IllegalArgumentException if the code is not &ge; 400 and &lt; 500.
      */
     public ClientErrorException(final int errorCode, final String msg, final Throwable cause) {
-        super(errorCode, msg, cause);
+        this(null, errorCode, msg, cause);
+    }
+
+    /**
+     * Creates a new exception for a tenant, a client error code, a detail message and a root cause.
+     *
+     * @param tenant The tenant that the exception occurred in the scope of or {@code null} if unknown.
+     * @param errorCode The code representing the erroneous outcome.
+     * @param msg The detail message.
+     * @param cause The root cause.
+     * @throws IllegalArgumentException if the code is not &ge; 400 and &lt; 500.
+     */
+    public ClientErrorException(final String tenant, final int errorCode, final String msg, final Throwable cause) {
+        super(tenant, errorCode, msg, cause);
         if (errorCode < 400 || errorCode >= 500) {
             throw new IllegalArgumentException("client error code must be >= 400 and < 500");
         }
     }
-
 }