+++ client/src/main/java/org/eclipse/hono/client/ServerErrorException.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2018 Contributors to the Eclipse Foundation
+ * Copyright (c) 2016, 2020 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -28,7 +28,18 @@ public class ServerErrorException extends ServiceInvocationException {
      * @throws IllegalArgumentException if the code is not &ge; 500 and &lt; 600.
      */
     public ServerErrorException(final int errorCode) {
-        this(errorCode, null, null);
+        this(null, errorCode, null, null);
+    }
+
+    /**
+     * Creates a new exception for a tenant and a server error code.
+     *
+     * @param tenant The tenant that the exception occurred in the scope of or {@code null} if unknown.
+     * @param errorCode The code representing the error that occurred.
+     * @throws IllegalArgumentException if the code is not &ge; 500 and &lt; 600.
+     */
+    public ServerErrorException(final String tenant, final int errorCode) {
+        this(tenant, errorCode, null, null);
     }
 
     /**
@@ -39,7 +50,19 @@ public class ServerErrorException extends ServiceInvocationException {
      * @throws IllegalArgumentException if the code is not &ge; 500 and &lt; 600.
      */
     public ServerErrorException(final int errorCode, final String msg) {
-        this(errorCode, msg, null);
+        this(null, errorCode, msg, null);
+    }
+
+    /**
+     * Creates a new exception for a tenant, a server error code and a detail message.
+     *
+     * @param tenant The tenant that the exception occurred in the scope of or {@code null} if unknown.
+     * @param errorCode The code representing the error that occurred.
+     * @param msg The detail message.
+     * @throws IllegalArgumentException if the code is not &ge; 500 and &lt; 600.
+     */
+    public ServerErrorException(final String tenant, final int errorCode, final String msg) {
+        this(tenant, errorCode, msg, null);
     }
 
     /**
@@ -50,7 +73,19 @@ public class ServerErrorException extends ServiceInvocationException {
      * @throws IllegalArgumentException if the code is not &ge; 500 and &lt; 600.
      */
     public ServerErrorException(final int errorCode, final Throwable cause) {
-        this(errorCode, null, cause);
+        this(null, errorCode, null, cause);
+    }
+
+    /**
+     * Creates a new exception for a tenant, a server error code and a root cause.
+     *
+     * @param tenant The tenant that the exception occurred in the scope of or {@code null} if unknown.
+     * @param errorCode The code representing the error that occurred.
+     * @param cause The root cause.
+     * @throws IllegalArgumentException if the code is not &ge; 500 and &lt; 600.
+     */
+    public ServerErrorException(final String tenant, final int errorCode, final Throwable cause) {
+        this(tenant, errorCode, null, cause);
     }
 
     /**
@@ -62,7 +97,20 @@ public class ServerErrorException extends ServiceInvocationException {
      * @throws IllegalArgumentException if the code is not &ge; 500 and &lt; 600.
      */
     public ServerErrorException(final int errorCode, final String msg, final Throwable cause) {
-        super(errorCode, msg, cause);
+        this(null, errorCode, msg, cause);
+    }
+
+    /**
+     * Creates a new exception for a tenant, a server error code, a detail message and a root cause.
+     *
+     * @param tenant The tenant that the exception occurred in the scope of or {@code null} if unknown.
+     * @param errorCode The code representing the error that occurred.
+     * @param msg The detail message.
+     * @param cause The root cause.
+     * @throws IllegalArgumentException if the code is not &ge; 500 and &lt; 600.
+     */
+    public ServerErrorException(final String tenant, final int errorCode, final String msg, final Throwable cause) {
+        super(tenant, errorCode, msg, cause);
         if (errorCode < 500 || errorCode >= 600) {
             throw new IllegalArgumentException("client error code must be >= 500 and < 600");
         }